requirements-dev.txt
ruff>=0.6.0
pytest>=8.0.0
pytest-homeassistant-custom-component>=0.13.0

